Update consul-leaders module
* use the new ASG, ssh-sg and cluster-network modules
* add missing AMI
* whitespace / ssh_pubkey variable cleanup
* other updates to consul-leaders module
* Drop consul-leader-sg from consul-leaders module

Rather than create this security group in the module, we will
provide the list of security groups to the consul leaders module.
This means the module is not responsible for the security group,
but the "caller" is. Thus, we simplify variables and swap for a
pass-through (leader_security_group_ids accepts string-lists).