set consul agent retry interval to 1s, in generic agent init
This is added as a variable in init, so it can be tuned a little more
easily, but the interval defaults to 1s.

With a short, 1 second delay, this speeds up the time it takes the agent
to joins the consul cluster. This speed up is needed, in part, because
the agent uses DNS on AWS to find the leaders, and that DNS points to
all
possible leader IPs, not the 3 IPs that are actually active (so the
agent
needs to iterate over 30 or so IPs to find 3). This sounds terrible, but
has been acceptable (it's reliable and works, albeit a little slow), and
the solution needs time/attention