add TF module that creates an IAM user and policy for accessing S3
This is different from the (existing) snapshots-bucket module in that
the caller is expected to create the S3 bucket, and this module _only_
creates the IAM user + policies. This module is intended for creating
multiple users (via separate calls to the module) for one bucket, rather
than coupling one IAM user to one bucket.