single-node-asg: add new module
This is a high-level module that incorporates two primary resources:

    * an autoscaling group with one node
    * an EBS volume with an IAM policy for the ASG
    * the init shell for the node to attach that EBS volume on init

This is a generic pattern that can be reused for prometheus, consul,
nomad, and other types of "core services" that have some state and
which run well having multiple leaders. EBS is bound by AZ, so this
module plays well with the various aspects of this pattern.