ec2-nat-instance: refactor the module for flexibility
This refactors the module in a few significant ways, though leaves the
core of the module and what it provides. The goal is to make this module
a bit more flexible for the user. Here is a rough list of changes:

* drop the security groups, let the caller provide a list
* refactor node init to use the init-snippet modules
* set the hostname during init
* fixup use of `extra_tags` so that is more useful
* drop the route table and subnet associations (let the caller provide)
* fixup variables and outputs to fit the details above
* rename `availability_zone` variable to `az` (match `azs` elsewhere)
* rename `aws_key_pair` variable to `key_name` (match other modules)

Also note that as of right now, changes to `user_data` are _ignored_,
and Terraform won't auto-destroy the resource on you.. the instance must
be tainted explicitly. Also note that while the module has been tested,
the instance init _has not been tested_, and may need a fixup commit
later. Not everything in life is perfect..