ec2-nat-instance: fixup init script
if you don't use echo '...' > /path/to/script, as part of cloud-init,
bash will attempt to process $s as a variable (even though it is not
set), and the script will not end up with $s in it. With echo '...',
we're able to retain the contents of the script unmodified by bash.