ec2-fixed-ip-auto-recover-instances: Create N instances in M subnets
the number of instances created is now dependent first on the number
of `private_ips` provided, but if that is an empty list, then we use
the number of `subnet_ids`. This allows a module user to create N
instances in M subnets, or N instances in N subnets. The number of
`private_ips` takes presedence, but this change means the module
supports a user who does not want to provide the IP address, and instead
lets AWS set the private IP for the instance.