ex/kube-stack-private: use `extra_tags` for changes in Kubernetes
Use `extra_tags` to add the `kubernetes.io/cluster/${var.name}` tag
to the VPC, subnets, route tables, NAT gateway, and any other resources
in the VPC Scenario 2 module that are using `extra_tags`.

Note that this coincides with updates to the `kube-stack` module, as
the tags are needed both by tools like kops as well as kubernetes
components (for load balancers, etc).