nat-gateways: use element() instead of var.foo[] syntax
When deploying the `vpc-scenario-2` example, `nat-gateways` would fail
with annoying and cryptic/difficult to debug errors about the array
indexing not matching what's in the array.

IDK why `element()` is better in this case, but a bunch of testing
confirmed this as a fix for the immediate problem. I will keep
investigating this situation, but this should address #131.

I haven't tested with other versions of Terraform, so IDK if this is
limited to 0.11.7 or is present in other versions.