asg: remove 'desired_capacity' and add 'termination_policies' input
When the capacity of an ASG should "float" (e.g. when using autoscaling
policies or schedules), setting the 'desired_capacity' in the ASG is
undesirable as it means anytime the TF is applied the capacity gets
reset.

Unfortunately it is not possible to make 'desired_capacity' optional, so
opted for removing it entirely.  'min_size' and 'max_size' can be used to
control the size instead.

Also added 'termination_policies' input, which is useful when autoscaling
to make sure the right instances are terminated when scaling down.

This change also required updating other modules that use the 'asg' module,
thus also removing desired capacity inputs from them.