add modules for credstash reader/writer grants
These two modules are based on the existing credstash-grant module,
but replaces the grant.sh script and local_exec provisioner that was
necessary to use in that module with the aws_kms_grant resource that
terraform now provides.

The one module is now split into two for directness in use and for
the simplicity of variables.