New function: single-node-asg module supports binding EIP by itself.
Since it is single node, binding an EIP to the instance is possible. And
it eases other things since the public interface is constant.

Add assign_eip variable to single-node-asg. If turns it on, an EIP will
be allocated, and assocated with the instance.

Scope VPC is specified in case the account does not have a default VPC.