Refine DLM lifecycle modules.
The original design has implict dependencies between `dlm-lifecycle-iam-role` and `dlm-lifecycle-policy` on IAM role. But TF currently does not support module dependency, and it does not get the dependency when a module is depends on another resource's `name` output.

So merged the two modules to avoid applying error that "the role does not exist".